Add borrower and lender association to user item.
The item is singular but it has 2 owners, once an item is borrowed its present location
or temporary owner is the borrower but technically it still belongs to the lender.
It is one item but at the same moment its tied to 2 people who are both from the user class.