Express relationship between users and items
Two key distinctions between the types of users we have here, lenders and borrowers have been
expressed in the user model
It is now possible to do 'User.borrowed_items to see all the borrowed items.
This code 'User.items' will list all the items that belong to a user since by default they
belong to a user. But to be able to create a list of lent items a scope will be used.